Rust enablement for LoongArch
On Thu, Jan 11, 2024 at 4:16 AM WANG Rui <wangrui@loongson.cn> wrote:
>
> 6.8

Thanks, that is quite expedited...

Please `rustfmt` the code to avoid failing the `rustfmtcheck` target.
I see you already created the 6.8 tag but I have not seen it in
linux-next yet and you did not send the PR, so it would be ideal if
you could fix this before it goes to Linus.